Crate foundry_fork_db

foundry-fork-db

License
Licensed under either of Apache License, Version 2.0 or MIT license at your option.
Unless you explicitly state otherwise, any contribution intentionally submitted for inclusion in these crates by you, as defined in the Apache-2.0 license, shall be dual licensed as above, without any additional terms or conditions.

Re-exports

pub use backend::BackendHandler;

pub use backend::SharedBackend;

pub use cache::BlockchainDb;

pub use error::DatabaseError;

pub use error::DatabaseResult;

Modules

backend

Smart caching and deduplication of requests when using a forking provider

cache

Cache related abstraction

error