Module foundry_fork_db::cache

Cache related abstraction

Structs

BlockchainDb

A shareable Block database

BlockchainDbMeta

relevant identifying markers in the context of BlockchainDb

FlushJsonBlockCacheDB

A type that flushes a JsonBlockCacheDB on drop

JsonBlockCacheDB

A DB that stores the cached content in a json file

JsonBlockCacheData

The Data the JsonBlockCacheDB can read and flush

MemDb

In Memory cache containing all fetched accounts and storage slots and their values from RPC

Type Aliases

StorageInfo